created controller models folder and moved respective methods to it
