+++ models/user.js
@@ -1,0 +1,10 @@
+const mongoose = require("mongoose");
+
+const userSchema = mongoose.Schema({
+  firstName: String,
+  lastName: String,
+  email: String,
+  password: String,
+});
+
+module.exports = mongoose.model("users", userSchema);
